Make vulkan_memory_allocator build with the MacOSX11.0 SDK.
The 11.0 SDK adds a declaration of aligned_malloc() that's available
on macOS 10.15 and later. vulkan_memory_allocator also defines a
symbol with the same name, which causes problems.

As fix, rename the vulkan_memory_allocator version to
vma_aligned_malloc(). Use this opportunity to use the same mechanism on
Windows (a no-op), and, while here, call the system aligned_malloc()
on macOS 10.15 from there too.

Bug: chromium:1098741
Change-Id: I346c7e1e3a1d6efba24997bd8612f5079434117c